+++ yasa/tests/test_others.py
@@ -1,20 +1,86 @@
 """
 Test the functions in the yasa/others.py file.
 """
+import mne
 import unittest
 import numpy as np
 from itertools import product
 from mne.filter import filter_data
-from yasa.others import (moving_transform, trimbothstd, _zerocrossings,
-                         sliding_window, get_centered_indices)
+
+from yasa.hypno import hypno_str_to_int, hypno_upsample_to_data
+from yasa.others import (moving_transform, trimbothstd, get_centered_indices,
+                         sliding_window, _merge_close, _zerocrossings,
+                         _index_to_events)
 
 # Load data
 data = np.loadtxt('notebooks/data_N2_spindles_15sec_200Hz.txt')
 sf = 200
 data_sigma = filter_data(data, sf, 12, 15, method='fir', verbose=0)
 
+# Load a full recording and its hypnogram
+file_full = np.load('notebooks/data_full_6hrs_100Hz_Cz+Fz+Pz.npz')
+data_full = file_full.get('data')
+chan_full = file_full.get('chan')
+sf_full = 100
+hypno_full = np.load('notebooks/data_full_6hrs_100Hz_hypno.npz').get('hypno')
 
-class TestStringMethods(unittest.TestCase):
+# Using MNE
+data_mne = mne.io.read_raw_fif('notebooks/sub-02_mne_raw.fif', preload=True,
+                               verbose=0)
+data_mne.pick_types(eeg=True)
+data_mne_single = data_mne.copy().pick_channels(['F3'])
+hypno_mne = np.loadtxt('notebooks/sub-02_hypno_30s.txt', dtype=str)
+hypno_mne = hypno_str_to_int(hypno_mne)
+hypno_mne = hypno_upsample_to_data(hypno=hypno_mne, sf_hypno=(1 / 30),
+                                   data=data_mne)
+
+
+class TestOthers(unittest.TestCase):
+
+    def test_index_to_events(self):
+        """Test functions _index_to_events"""
+        a = np.array([[3, 6], [8, 12], [14, 20]])
+        good = [3, 4, 5, 6, 8, 9, 10, 11, 12, 14, 15, 16, 17, 18, 19, 20]
+        out = _index_to_events(a)
+        np.testing.assert_equal(good, out)
+
+    # def test_get_sync_events(self):
+    #     """Test functions get_sync_events"""
+    #     sw = sw_detect(data_full, sf_full, chan_full)
+    #     df_sync = sw.get_sync_events(time_before=0.2, center='PosPeak')
+    #     df_sync2 = get_sync_events(data_full, sf_full, sw.summary(),
+    #                                time_before=0.2, center='PosPeak')
+    #     pd.testing.assert_frame_equal(df_sync, df_sync2, check_dtype=False)
+    #     assert df_sync['Channel'].nunique() == 3
+    #     # MNE single-channel
+    #     sw = sw.summary()
+    #     sw_c = sw[sw['Channel'] == sw.at[0, 'Channel']].iloc[:, :-2]
+    #     df_sync = get_sync_events(data_full[0, :], sf_full, sw_c,
+    #                               center='PosPeak', time_before=0,
+    #                               time_after=2)
+    #     assert df_sync.shape[1] == 3
+    #     # MNE single channel
+    #     sw = sw_detect(data_mne_single, amp_neg=(20, 300),
+    #                    amp_ptp=(60, 500))
+    #     df_sync = sw.get_sync_events()
+    #     df_sync2 = get_sync_events(data_mne_single, detection=sw.summary())
+    #     pd.testing.assert_frame_equal(df_sync, df_sync2, check_dtype=False)
+    #
+    #     # MNE multi channel
+    #     sw = sw_detect(data_mne, amp_neg=(20, 300), amp_ptp=(60, 500))
+    #     df_sync = sw.get_sync_events()
+    #     df_sync2 = get_sync_events(data_mne, detection=sw.summary())
+    #     pd.testing.assert_frame_equal(df_sync, df_sync2, check_dtype=False)
+    #     assert df_sync['Channel'].nunique() == 6
+
+    def test_merge_close(self):
+        """Test functions _merge_close"""
+        a = np.array([4, 5, 6, 7, 10, 11, 12, 13, 20, 21, 22, 100, 102])
+        good = np.array([4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16,
+                         17, 18, 19, 20, 21, 22, 100, 101, 102])
+        # Events that are less than 100 ms apart (i.e. 10 points at 100 Hz sf)
+        out = _merge_close(a, 100, 100)
+        np.testing.assert_equal(good, out)
 
     def test_moving_transform(self):
         """Test moving_transform"""
@@ -23,7 +89,6 @@ class TestStringMethods(unittest.TestCase):
         interp = [False, True]
         win = [.3, .5]
         step = [0, .5]
-
         prod_args = product(win, step, method, interp)
 
         for i, (w, s, m, i) in enumerate(prod_args):
@@ -37,7 +102,9 @@ class TestStringMethods(unittest.TestCase):
         """Test function trimbothstd
         """
         x = [4, 5, 7, 0, 18, 6, 7, 8, 9, 10]
+        y = np.random.normal(size=(10, 100))
         assert trimbothstd(x) < np.std(x, ddof=1)
+        assert (trimbothstd(y) < np.std(y, ddof=1, axis=-1)).all()
 
     def test_zerocrossings(self):
         """Test _zerocrossings